addressed jest issues
