+++ pages/projects.tsx
@@ -1,9 +1,8 @@
-import { Octokit } from '@octokit/core';
 import { GetStaticProps } from 'next';
 import React from 'react';
 import Layout from '../components/Layout';
 import ProjectCard from '../components/ProjectCard';
-import { Project } from '../util';
+import { getProjects, Project } from '../util';
 
 interface ProjectsProps {
   projects: Project[];
@@ -37,29 +36,7 @@ function Projects({ projects }: ProjectsProps): JSX.Element {
 export default Projects;
 
 export const getStaticProps: GetStaticProps<ProjectsProps> = async () => {
-  const octokit = new Octokit();
-  const projectsResponse = await octokit.request('GET /orgs/{org}/repos', {
-    org: 'uclaacm',
-    per_page: 100,
-  });
-
-  const sortedData = projectsResponse.data.sort(
-    (a, b) =>
-      new Date(b.updated_at).getTime() - new Date(a.updated_at).getTime(),
-  );
-  const projects = sortedData.map(
-    (repo) =>
-      ({
-        name: repo.name,
-        description: repo.description,
-        link: repo.homepage || null,
-        repo: repo.html_url,
-        lang: repo.language,
-        topics: repo.topics,
-        image: repo.owner.avatar_url,
-        alt: repo.owner.name || null,
-      } as Project),
-  );
+  const projects = await getProjects();
 
   return {
     props: {
